Move check to separate method
The method will be reused by docker and other pools.
