+++ src/DataIsland.SqlServer/OneSqlServer.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.Data.SqlClient;
 
 namespace DataIsland.SqlServer;
 
@@ -27,35 +26,9 @@ internal sealed class OneSqlServer : IComponentPool<SqlServerComponent, SqlServe
         if (componentSpec.ComponentType != typeof(SqlServerComponent))
             throw new InvalidOperationException("Incorrect component type.");
 
-        if (componentSpec.Collation is not null)
-        {
-            using var connection = new SqlConnection(_component.ConnectionString);
-            await connection.OpenAsync();
-            var command = connection.CreateCommand();
-            command.CommandText = "SELECT SERVERPROPERTY('Collation')";
-            var collation = (string?)await command.ExecuteScalarAsync();
-            connection.Close();
-            if (collation != componentSpec.Collation)
-                throw new InvalidOperationException($"Unable to find a server with collation {componentSpec.Collation}.");
-        }
-
-        if (componentSpec.ClrEnabled is not null)
-        {
-            var desiredClrEnabled = componentSpec.ClrEnabled.Value;
-            using var connection = new SqlConnection(_component.ConnectionString);
-            await connection.OpenAsync();
-            var command = connection.CreateCommand();
-            command.CommandText = """
-                                  DECLARE @T TABLE ([name] nvarchar(35), [minimum] int, [maximum] int, [config_value] int, [run_value] int);
-                                  INSERT @T EXEC sp_configure 'clr enabled';
-                                  SELECT run_value FROM @T;
-                                  """;
-            var clrConfigValue = (int?)await command.ExecuteScalarAsync();
-            var actualClrEnabled = clrConfigValue is not null && clrConfigValue.Value != 0;
-            connection.Close();
-            if (actualClrEnabled != desiredClrEnabled)
-                throw new InvalidOperationException($"Unable to find a server with clr {(desiredClrEnabled ? "enabled" : "disabled")}.");
-        }
+        var error = await SqlServerSpecChecker.CheckAsync(componentSpec, _component.ConnectionString);
+        if (error is not null)
+            throw new InvalidOperationException(error);
 
         return new Dictionary<string, SqlServerComponent>
         {
